requirements/_base.txt
colorama
crispy-tailwind
apscheduler
cryptography
Django
django-admin-rangefilter
django-components
django-crispy-forms
django-extensions
django-filter
django-import-export
django-log-viewer
django-maintenance-mode
django-redis
django-session-timeout
django-guardian
django-widget-tweaks
djangorestframework
drf-standardized-errors
django-easy-audit
gunicorn
hiredis
django-valkey
Pillow
prettytable
psycopg[binary]
python-decouple
python-magic
python-webpack-boilerplate
PyYAML
redis
requests
pandas
numpy
stringcase
sentry-sdk[django]
Whitenoise
